requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b9
aiohttp==3.8.5
aioselenium==0.0.1
aiosignal==1.3.1
alembic==1.11.2
annotated-types==0.5.0
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
environs==9.5.0
exceptiongroup==1.1.2
fake-useragent==1.2.1
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
HTMLParser==0.0.2
idna==3.4
loguru==0.7.0
magic-filter==1.0.11
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
outcome==1.2.0
packaging==23.1
psycopg2-binary==2.9.6
pydantic==2.1.1
pydantic_core==2.4.0
PySocks==1.7.1
python-dotenv==1.0.0
pytz==2023.3
redis==4.6.0
requests==2.31.0
selectolax==0.3.16
selenium==4.11.2
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.19
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
urllib3==2.0.4
useragent==0.1.1
webdriver-manager==4.0.0
wsproto==1.2.0
yarl==1.9.2
